general cleanup before merge
